Move the object reordering to the layer abstract class
