+++ packages/core/src/renderers/webgl_renderer.ts
@@ -16,7 +16,6 @@ import { Camera } from "../objects/cameras/camera";
 
 import { mat4, vec2, vec3 } from "gl-matrix";
 import { Frustum } from "../math/frustum";
-import { VolumeLayer } from "@/layers/volume_layer";
 
 // The library's coordinate system is left-handed.
 // With the default camera, the standard basis vectors should
@@ -118,7 +117,8 @@ export class WebGLRenderer extends Renderer {
     if (layer.type === "VolumeLayer") {
       this.state_.setCullFaceMode("back");
       this.state_.setDepthTesting(false);
-      (layer as VolumeLayer).reorderObjects(camera, "back-to-front");
+      this.state_.setDepthMask(false);
+      layer.reorderObjects(camera, "front-to-back");
     }
 
     layer.objects.forEach((object, i) => {
@@ -131,6 +131,7 @@ export class WebGLRenderer extends Renderer {
     if (layer.type === "VolumeLayer") {
       this.state_.setCullFaceMode("disabled");
       this.state_.setDepthTesting(true);
+      this.state_.setDepthMask(true);
     }
   }
